pyproject.toml
[tool.poetry]
name = "michael-bnb-backend"
version = "0.1.0"
description = ""
authors = ["michael <michael.yc.cho@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1"
Pillow = "^9.2.0"
PyJWT = "^2.4.0"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.2"
django-model-utils = "^4.3.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
